Change print_num_base argument data type from unsigned int to unsigned long

+++ num_functions.c
@@ -38,7 +38,7 @@ int print_int(long int n)
  * Return: Number of characters printed
  */
 
-int print_num_base(unsigned int num, int base)
+int print_num_base(unsigned long num, int base)
 {
 	char buf[BUF_SIZE];
 	int index = BUF_SIZE - 1;
